split items
