+++ projects/Mallard/src/components/front/helpers/helpers.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent } from 'react'
-import { PropTypes } from '../items/item'
+import { PropTypes } from '../items/items'
 import { FlatList } from 'react-native'
 import { Animated } from 'react-native'
 import { metrics } from 'src/theme/spacing'
